+++ example/send_gmail.dart
@@ -4,8 +4,8 @@ import 'package:mailer/mailer.dart';
 
 /// Test mailer by sending email to yourself
 main(List<String> args) {
-  if (args.length != 2) {
-    print('Usage: send_gmail <gmail_username> <gmail_password>');
+  if (args.length < 2 || args.length > 3) {
+    print('Usage: send_gmail <username> <password> [<file-to-attach>]');
     print('');
     print('If you have Google\'s "app specific passwords" enabled,');
     print('you need to use one of those for the password here.');
@@ -16,6 +16,14 @@ main(List<String> args) {
   String password = args[1];
   if (username.endsWith('@gmail.com')) {
     username = username.substring(0, username.length - 10);
+  }
+  File attachFile;
+  if (args.length > 2) {
+    attachFile = new File(args[2]);
+    if (!attachFile.existsSync()) {
+      print('Failed to find file to attach: ${attachFile.path}');
+      exit(1);
+    }
   }
 
   // If you want to use an arbitrary SMTP server, go with `new SmtpOptions()`.
@@ -31,9 +39,12 @@ main(List<String> args) {
   var envelope = new Envelope()
     ..from = '$username@gmail.com'
     ..recipients.add('$username@gmail.com')
-    ..subject = 'Testing the Dart Mailer library'
+    ..subject = 'Test Dart Mailer library :: ${new DateTime.now()}'
     ..text = 'This is the plain text'
     ..html = '<h1>Test</h1><p>Hey! Here\'s some HTML content</p>';
+  if (attachFile != null) {
+    envelope.attachments.add(new Attachment(file: attachFile));
+  }
 
   // Email it.
   emailTransport