+++ lib/src/envelope.dart
@@ -93,8 +93,8 @@ class Envelope {
         var filename = basename(attachment.file.path);
 
         return attachment.file.readAsBytes().then((bytes) {
-          // Create a chunk'd (76 chars per line) base64 string.
-          var contents = CryptoUtils.bytesToBase64(bytes, addLineSeparator: true);
+          // Chunk'd (76 chars per line) base64 string, separated by "\r\n".
+          var contents = chunkEncodedBytes(BASE64.encode(bytes));
 
           buffer.write('--$boundary\n');
           buffer.write('Content-Type: ${_getMimeType(attachment.file.path)}; name="$filename"\n');
@@ -109,6 +109,23 @@ class Envelope {
       });
     });
   }
+}
+
+/// Consume the encode bytes (no line separators),
+/// and produce a chunk'd (76 chars per line) string, separated by "\r\n".
+chunkEncodedBytes(String encoded) {
+  if (encoded == null) return null;
+  var chunked = new StringBuffer();
+  int start = 0;
+  int end = encoded.length;
+  do {
+    int next = start + 76;
+    if (next > end) next = end;
+    chunked.write(encoded.substring(start, next));
+    chunked.write('\r\n');
+    start = next;
+  } while (start < encoded.length);
+  return chunked.toString();
 }
 
 String _getMimeType(String path) {